src/api_gateway/requirements.txt
fastapi
uvicorn
kafka-python==2.0.2
websockets
